Add ZX-81 "Shoulder of Giants" ROM by Geoff Wearmouth
This is a customized ZX81 ROM variant by Geoff Wearmouth which
uses space-saving techniques to make way for Newton's square root
calculation, improved decimal number input and consistent output to
the screen and printer of floating point numbers.

The first commit is the original file from the author. The next commits
will merge with the ZX-81 rom source.